Some more migrations; refs #1354
